Separate time spent list and details into two components
